[Feature] Unify Components (#159)
* Components: Unify forms

add prepare-from for remove same components (like treasury-increase-form.vue, savings-deposit-form.vue etc.)
add review-form for the same purpose
start remove same components

* Components: Unify forms

apply new components for decrease, increase and claim&burn
fix description for decrease from (later use text from increase)

* Components: Unify forms

final refactoring:
- move 'full-page-loader' to forms directory and rename to 'loader-form'
- add index.ts files for form directory
- remove index.ts from loader-form directory (unusable)
- update import lines
- rename some types

* Components: Unify forms

remove legacy components

* Components: Unify forms - fixes

Fixes after code-review

* Components: Unify forms - savings

- apply new unify components to savings pages
- remove legacy components
- fix treasury wrapper components
- add new prop (has-select-modal) to prepare-form

* Components: fix

Remove template string from review-form

* Components: fix

Fix after code-review
remove legacy and unused filed from token

* Components: fix

Fix naming of computed filed

* Components: fix

change button to action-button
