+++ src/components/buttons/action-button.vue
@@ -3,7 +3,8 @@
     :class="[classes]"
     :disabled="disabled"
     :style="customStyle"
-    @click.prevent.stop="handleClick"
+    :type="type"
+    @click="handleClick"
   >
     <template v-if="text">
       {{ text }}
@@ -35,6 +36,14 @@ export default Vue.extend({
     customStyle: {
       type: Object as PropType<CssProperties>,
       default: undefined
+    },
+    type: {
+      type: String,
+      default: 'button'
+    },
+    propagateOriginalEvent: {
+      type: Boolean,
+      default: false
     }
   },
   computed: {
@@ -43,7 +52,12 @@ export default Vue.extend({
     }
   },
   methods: {
-    handleClick(): void {
+    handleClick(event: MouseEvent): void {
+      if (!this.propagateOriginalEvent) {
+        event.preventDefault();
+        event.stopPropagation();
+      }
+
       this.$emit('button-click');
     }
   }
