--- src/components/controls/full-page-form-loader/index.ts
@@ -1,2 +1,0 @@
-export { default as FullPageFormLoader } from './full-page-form-loader.vue';
-export type { Step } from './types';
