--- src/components/controls/full-page-form-loader/types.ts
@@ -1,7 +1,0 @@
-export type Step = 'Confirm' | 'Process' | 'Success' | 'Reverted';
-
-export type StepData = {
-  id: Step;
-  title: string;
-  subtitle: string;
-};
