+++ src/components/forms/index.ts
@@ -1,0 +1,6 @@
+export { default as PrepareForm } from './prepare-form/prepare-form.vue';
+export { default as ReviewForm } from './review-form.vue';
+export { default as LoaderForm } from './loader-form/loader-form.vue';
+
+export type { LoaderStep } from './loader-form/types';
+export type { InputMode } from './prepare-form/types';
