+++ src/components/forms/loader-form/types.ts
@@ -1,0 +1,7 @@
+export type LoaderStep = 'Confirm' | 'Process' | 'Success' | 'Reverted';
+
+export type StepData = {
+  id: LoaderStep;
+  title: string;
+  subtitle: string;
+};
