+++ src/components/forms/prepare-form/types.ts
@@ -1,0 +1,1 @@
+export type InputMode = 'NATIVE' | 'TOKEN';
