+++ src/components/savings/index.ts
@@ -4,8 +4,3 @@ export { default as SavingsYearlyChartWrapper } from './savings-yearly-chart-wra
 export { default as SavingsLeftRail } from './savings-left-rail.vue';
 export { default as SavingsEmpty } from './savings-empty.vue';
 export { default as SavingsManage } from './savings-manage.vue';
-export { default as SavingsDepositForm } from './savings-deposit-form.vue';
-export { default as SavingsWithdrawForm } from './savings-withdraw-form.vue';
-export { default as SavingsWithdrawReview } from './savings-withdraw-review.vue';
-export { default as SavingsDepositReview } from './savings-deposit-review.vue';
-export { default as SavingsFormLoader } from '../controls/full-page-form-loader/full-page-form-loader.vue';
