+++ src/components/treasury/treasury-powercard-empty.vue
@@ -50,7 +50,7 @@
         </div>
       </div>
     </template>
-    <full-page-form-loader v-else :step="txStep" />
+    <loader-form v-else :step="txStep" />
   </secondary-page>
 </template>
 
@@ -65,15 +65,13 @@ import { stakePowercardCompound } from '@/wallet/actions/treasury/powercard/stak
 import { estimateStakePowercardCompound } from '@/wallet/actions/treasury/powercard/stakeEstimate';
 
 import { ActionButton } from '@/components/buttons';
-import { FullPageFormLoader } from '@/components/controls/full-page-form-loader';
+import { LoaderForm, LoaderStep } from '@/components/forms';
 import { PictureDescriptor } from '@/components/html5';
 import CustomPicture from '@/components/html5/custom-picture.vue';
 import {
   SecondaryPage,
   SecondaryPageSimpleTitle
 } from '@/components/layout/secondary-page';
-
-import { Step } from '../controls/form-loader';
 
 export default Vue.extend({
   name: 'TreasuryPowercardEmpty',
@@ -82,11 +80,11 @@ export default Vue.extend({
     ActionButton,
     SecondaryPageSimpleTitle,
     SecondaryPage,
-    FullPageFormLoader
+    LoaderForm
   },
   data() {
     return {
-      txStep: undefined as Step | undefined,
+      txStep: undefined as LoaderStep | undefined,
       actionError: undefined as string | undefined,
       powercard: {
         alt: this.$t('treasury.lblSmartTreasury'),
@@ -148,7 +146,7 @@ export default Vue.extend({
           this.currentAddress,
           resp.actionGasLimit,
           resp.approveGasLimit,
-          async (step: Step) => {
+          async (step: LoaderStep) => {
             this.txStep = step;
           }
         );
