+++ src/components/treasury/treasury-powercard-manage.vue
@@ -1,6 +1,6 @@
 <template>
   <secondary-page has-back-button hide-title @back="handleBack">
-    <template v-if="txStep === undefined">
+    <template v-if="transactionStep === undefined">
       <div>
         <secondary-page-simple-title
           class="page-title max-width"
@@ -51,7 +51,7 @@
         </div>
       </div>
     </template>
-    <full-page-form-loader v-else :step="txStep" />
+    <loader-form v-else :step="transactionStep" />
   </secondary-page>
 </template>
 
@@ -70,7 +70,7 @@ import { unstakePowercardCompound } from '@/wallet/actions/treasury/powercard/un
 import { estimateUnstakePowercardCompound } from '@/wallet/actions/treasury/powercard/unstakeEstimate';
 
 import ActionButton from '@/components/buttons/action-button.vue';
-import { FullPageFormLoader } from '@/components/controls/full-page-form-loader';
+import { LoaderForm, LoaderStep } from '@/components/forms';
 import { PictureDescriptor } from '@/components/html5';
 import CustomPicture from '@/components/html5/custom-picture.vue';
 import { ProgressLoader } from '@/components/layout';
@@ -78,8 +78,6 @@ import {
   SecondaryPage,
   SecondaryPageSimpleTitle
 } from '@/components/layout/secondary-page';
-
-import { Step } from '../controls/form-loader';
 
 export default Vue.extend({
   name: 'TreasuryPowercardManage',
@@ -89,12 +87,12 @@ export default Vue.extend({
     ProgressLoader,
     SecondaryPageSimpleTitle,
     SecondaryPage,
-    FullPageFormLoader
+    LoaderForm
   },
   data() {
     return {
       actionError: undefined as string | undefined,
-      txStep: undefined as Step | undefined,
+      transactionStep: undefined as LoaderStep | undefined,
       powercard: {
         alt: this.$t('treasury.lblSmartTreasury'),
         src: require('@/assets/images/Powercard@1x.png'),
@@ -177,7 +175,7 @@ export default Vue.extend({
         return;
       }
 
-      this.txStep = 'Confirm';
+      this.transactionStep = 'Confirm';
       try {
         await unstakePowercardCompound(
           this.networkInfo.network,
@@ -185,14 +183,14 @@ export default Vue.extend({
           this.currentAddress,
           resp.actionGasLimit,
           resp.approveGasLimit,
-          async (step: Step) => {
-            this.txStep = step;
+          async () => {
+            this.transactionStep = 'Process';
           }
         );
-        this.txStep = 'Success';
+        this.transactionStep = 'Success';
         this.updateWalletAfterTxn();
       } catch (err) {
-        this.txStep = 'Reverted';
+        this.transactionStep = 'Reverted';
         Sentry.captureException(err);
       }
     }
