+++ src/styles/_common.less
@@ -736,7 +736,8 @@ body {
   height: 40px;
   color: #fff;
   background: #fff;
-  background: linear-gradient(-135deg, transparent 50%, var(--bgColor) 0) no-repeat 100% ~'0 / 12px 12px',
+  background:
+ linear-gradient(-135deg, transparent 50%, var(--bgColor) 0) no-repeat 100% ~'0 / 12px 12px',
     linear-gradient(-135deg, transparent 9px, #fff 0);
   filter: drop-shadow(0 0 16px #ccc);
   border-radius: 3px;
@@ -813,7 +814,7 @@ body {
   margin-top: 40px;
 
   &.max-width,
-  .max-width{
+  .max-width {
     max-width: 480px;
   }
 
@@ -1026,7 +1027,6 @@ button > .loader-icon > img {
       font-weight: 700;
       font-size: 24px;
       line-height: 40px;
-
       position: absolute;
       left: -9999px;
       top: -9999px;
@@ -1054,6 +1054,8 @@ button > .loader-icon > img {
       font-family: 'Medium', sans-serif;
       font-weight: 500;
       color: rgba(60, 60, 67, 0.6);
+      font-size: 14px;
+      line-height: 22px;
     }
   }
