+++ src/styles/_savings.less
@@ -343,219 +343,6 @@
   }
 }
 
-.review__wrapper {
-  .title {
-    font-family: 'Bold', sans-serif;
-    font-weight: 700;
-    font-size: 40px;
-    line-height: 54px;
-  }
-
-  .arrow {
-    margin-top: 40px;
-    display: -webkit-box;
-    display: -webkit-flex;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-align: center;
-    -webkit-align-items: center;
-    -ms-flex-align: center;
-    align-items: center;
-
-    picture {
-      display: block;
-
-      img {
-        display: block;
-        width: 100%;
-      }
-    }
-
-    .item-coin {
-      width: 40px;
-      height: 40px;
-
-      img {
-        display: block;
-        width: 100%;
-        border-radius: 50%;
-      }
-
-      .img-stub {
-        border-radius: 50%;
-        width: 38px;
-        height: 38px;
-        background-color: #f1f1f1;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        box-shadow: 0 0 16px rgba(0, 0, 0, 50%);
-
-        span {
-          font-family: 'Bold', sans-serif;
-          font-weight: 700;
-          font-size: 9px;
-          color: #000;
-          text-align: center;
-          line-height: 16px;
-        }
-      }
-    }
-
-    .item-image {
-      width: 54px;
-      height: 40px;
-    }
-
-    .item-arrow {
-      -webkit-transform: rotate(-90deg);
-      -ms-transform: rotate(-90deg);
-      transform: rotate(-90deg);
-
-      span {
-        display: block;
-        width: 17px;
-        height: 17px;
-        border-bottom: 2px solid rgba(0, 0, 0, 0.5);
-        border-right: 2px solid rgba(0, 0, 0, 0.5);
-        -webkit-transform: rotate(45deg) skew(6deg, 6deg);
-        -ms-transform: rotate(45deg) skew(6deg, 6deg);
-        transform: rotate(45deg) skew(6deg, 6deg);
-        border-radius: 1px;
-        -webkit-animation: arrow-animate 1s infinite;
-        animation: arrow-animate 1s infinite;
-      }
-
-      span:nth-child(1) {
-        -webkit-animation-delay: -0.3s;
-        animation-delay: -0.3s;
-      }
-
-      span:nth-child(2) {
-        margin: -4px 0 0;
-      }
-    }
-
-    .item {
-      margin-left: 40px;
-
-      .shadow {
-        filter: drop-shadow(0 4px 16px #ccc);
-      }
-    }
-
-    .item:first-child {
-      margin-left: 0;
-    }
-  }
-
-  .items {
-    margin-top: 40px;
-
-    .item {
-      margin-top: 40px;
-
-      h2 {
-        font-family: 'Medium', sans-serif;
-        font-weight: 500;
-        font-size: 16px;
-        line-height: 22px;
-      }
-
-      span {
-        display: block;
-        margin-top: 7px;
-        font-family: 'Medium', sans-serif;
-        font-weight: 500;
-        font-size: 16px;
-        line-height: 24px;
-        color: rgba(60, 60, 67, 0.6);
-      }
-    }
-
-    .item:first-child {
-      margin-top: 0;
-    }
-  }
-
-  .switch {
-    margin-top: 40px;
-    display: -webkit-box;
-    display: -webkit-flex;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-align: center;
-    -webkit-align-items: center;
-    -ms-flex-align: center;
-    align-items: center;
-
-    p {
-      font-family: 'Medium', sans-serif;
-      font-weight: 500;
-      font-size: 16px;
-      line-height: 22px;
-      margin-right: 108px;
-    }
-
-    .switch-button {
-      display: inline-block;
-      position: relative;
-      width: 56px;
-      height: 32px;
-      border-radius: 80px;
-      border: 1px solid #d8d8d8;
-      background: #fff;
-      -webkit-transition: background 0.28s cubic-bezier(0.4, 0, 0.2, 1);
-      -o-transition: background 0.28s cubic-bezier(0.4, 0, 0.2, 1);
-      transition: background 0.28s cubic-bezier(0.4, 0, 0.2, 1);
-      vertical-align: middle;
-      cursor: pointer;
-    }
-
-    .switch-button::before {
-      content: '';
-      position: absolute;
-      top: 2px;
-      left: 2px;
-      width: 26px;
-      height: 26px;
-      background: #fff;
-      border-radius: 50%;
-      -webkit-box-shadow: 0 1px 3px rgba(0, 0, 0, 0.5);
-      box-shadow: 0 1px 3px rgba(0, 0, 0, 0.5);
-      -webkit-transition: left 0.28s cubic-bezier(0.4, 0, 0.2, 1), background 0.28s cubic-bezier(0.4, 0, 0.2, 1),
-        -webkit-box-shadow 0.28s cubic-bezier(0.4, 0, 0.2, 1);
-      -o-transition: left 0.28s cubic-bezier(0.4, 0, 0.2, 1), background 0.28s cubic-bezier(0.4, 0, 0.2, 1),
-        box-shadow 0.28s cubic-bezier(0.4, 0, 0.2, 1);
-      transition: left 0.28s cubic-bezier(0.4, 0, 0.2, 1), background 0.28s cubic-bezier(0.4, 0, 0.2, 1),
-        box-shadow 0.28s cubic-bezier(0.4, 0, 0.2, 1), -webkit-box-shadow 0.28s cubic-bezier(0.4, 0, 0.2, 1);
-    }
-
-    .switch-button:active::before {
-      -webkit-box-shadow: 0 2px 8px rgba(0, 0, 0, 0.28), 0 0 0 20px rgba(128, 128, 128, 0.1);
-      box-shadow: 0 2px 8px rgba(0, 0, 0, 0.28), 0 0 0 20px rgba(128, 128, 128, 0.1);
-    }
-
-    input:checked + .switch-button {
-      background: #000;
-    }
-
-    input:checked + .switch-button::before {
-      left: 27px;
-      background: #fff;
-    }
-
-    input:checked + .switch-button:active::before {
-      -webkit-box-shadow: 0 2px 8px rgba(0, 0, 0, 0.28), 0 0 0 20px rgba(0, 150, 136, 0.2);
-      box-shadow: 0 2px 8px rgba(0, 0, 0, 0.28), 0 0 0 20px rgba(0, 150, 136, 0.2);
-    }
-  }
-
-  .button-active {
-    margin-top: 40px;
-  }
-}
-
 @-webkit-keyframes arrow-animate {
   0% {
     opacity: 0;
