+++ src/views/savings/savings-deposit-wrapper.vue
@@ -4,117 +4,546 @@
     hide-title
     @back="handleBack"
   >
-    <savings-deposit-form v-if="!isShowReview" @tx-review="handleTxReview" />
-    <savings-deposit-review
-      v-else-if="txStep === undefined"
-      :amount="amount"
+    <prepare-form
+      v-if="step === 'prepare'"
+      :asset="inputAsset"
+      has-select-modal
+      :header-description="
+        $t('treasury.decreaseBoost.txtDecreaseBoostPageDescription')
+      "
+      :header-title="$t('savings.deposit.lblDepositInSavings')"
+      :input-amount="inputAmount"
+      :input-amount-native="inputAmountNative"
+      :input-asset-heading="$t('savings.deposit.lblWhatDoWeDeposit')"
+      :input-mode="inputMode"
+      :is-loading="isLoading"
+      :is-processing="isProcessing"
+      :operation-description="$t('savings.deposit.txtYouCouldEarnInYear')"
+      :operation-title="estimatedAnnualEarning"
+      :output-asset-heading-text="$t('savings.deposit.lblAmountWeDepositIn')"
+      :selected-token-description="description"
+      :transfer-error="transferError"
+      @open-select-modal="handleOpenSelectModal"
+      @review-tx="handleTxReview"
+      @select-max-amount="handleSelectMaxAmount"
+      @toggle-input-mode="handleToggleInputMode"
+      @update-amount="handleUpdateAmount"
+    >
+      <template v-slot:swap-message>
+        <div
+          v-if="isSwapNeeded && formattedUSDCTotal && inputMode === 'TOKEN'"
+          class="form-swap"
+        >
+          <p>
+            {{ $t('forms.lblSwappingFor') }}
+            <custom-picture
+              :alt="$t('lblUSDcTokenAlt')"
+              class="token"
+              :sources="usdcPicture.sources"
+              :src="usdcPicture.src"
+            />
+            <span>{{ formattedUSDCTotal }}</span>
+          </p>
+        </div>
+      </template>
+    </prepare-form>
+    <review-form
+      v-else-if="step === 'review'"
+      :amount="inputAmount"
+      :button-text="$t('savings.deposit.lblDepositInSavings')"
       :estimated-gas-cost="estimatedGasCost"
+      :header-title="$t('savings.deposit.lblReviewYourDeposit')"
+      :image="savings"
+      :input-amount-native-title="$t('savings.deposit.lblAndTotalOf')"
+      :input-amount-title="$t('savings.deposit.lblAmountWeDepositIn')"
       :is-subsidized-enabled="isSubsidizedEnabled"
-      :native-amount="nativeAmount"
-      :token="token"
+      :native-amount="inputAmountNative"
+      :token="inputAsset"
       @tx-start="handleTxStart"
     />
-    <full-page-form-loader v-else :step="txStep" />
+    <loader-form v-else-if="step === 'loader'" :step="transactionStep" />
   </secondary-page>
 </template>
 
 <script lang="ts">
 import Vue from 'vue';
-import { mapActions, mapState } from 'vuex';
+import { mapActions, mapGetters, mapState } from 'vuex';
 
 import * as Sentry from '@sentry/vue';
+import BigNumber from 'bignumber.js';
 
-import { TransferData } from '@/services/0x/api';
+import {
+  getTransferData,
+  TransferData,
+  ZeroXSwapError
+} from '@/services/0x/api';
+import { mapError } from '@/services/0x/errors';
+import { Modal as ModalType } from '@/store/modules/modals/types';
 import { sameAddress } from '@/utils/address';
+import {
+  add,
+  convertAmountFromNativeValue,
+  convertNativeAmountFromAmount,
+  divide,
+  fromWei,
+  isZero,
+  multiply,
+  toWei
+} from '@/utils/bigmath';
+import { formatToNative } from '@/utils/format';
 import { depositCompound } from '@/wallet/actions/savings/deposit/deposit';
+import { estimateDepositCompound } from '@/wallet/actions/savings/deposit/depositEstimate';
+import {
+  calcTransactionFastNativePrice,
+  isSubsidizedAllowed
+} from '@/wallet/actions/subsidized';
+import { CompoundEstimateResponse } from '@/wallet/actions/types';
 import { getUSDCAssetData } from '@/wallet/references/data';
-import { SmallTokenInfoWithIcon, TokenWithBalance } from '@/wallet/types';
+import {
+  SmallToken,
+  SmallTokenInfoWithIcon,
+  TokenWithBalance
+} from '@/wallet/types';
 
-import { Step } from '@/components/controls/form-loader/types';
-import { FullPageFormLoader } from '@/components/controls/full-page-form-loader';
+import {
+  InputMode,
+  LoaderForm,
+  LoaderStep,
+  PrepareForm,
+  ReviewForm
+} from '@/components/forms';
+import { CustomPicture, PictureDescriptor } from '@/components/html5';
 import { SecondaryPage } from '@/components/layout/secondary-page';
-import { SavingsDepositForm, SavingsDepositReview } from '@/components/savings';
+
+type ProcessStep = 'prepare' | 'review' | 'loader';
 
 export default Vue.extend({
   name: 'SavingsDepositWrapper',
   components: {
-    FullPageFormLoader,
-    SecondaryPage,
-    SavingsDepositReview,
-    SavingsDepositForm
+    CustomPicture,
+    ReviewForm,
+    PrepareForm,
+    LoaderForm,
+    SecondaryPage
   },
   data() {
     return {
-      isShowReview: false,
-      txStep: undefined as Step | undefined,
+      //current
+      step: 'prepare' as ProcessStep,
+      transactionStep: undefined as LoaderStep | undefined,
+      savings: {
+        alt: this.$t('savings.lblSavings'),
+        src: require('@/assets/images/Savings@1x.png'),
+        sources: [
+          { src: require('@/assets/images/Savings@1x.png') },
+          {
+            variant: '2x',
+            src: require('@/assets/images/Savings@2x.png')
+          }
+        ],
+        webpSources: [
+          { src: require('@/assets/images/Savings@1x.webp') },
+          {
+            variant: '2x',
+            src: require('@/assets/images/Savings@2x.webp')
+          }
+        ]
+      } as PictureDescriptor,
+      usdcPicture: {
+        src: require('@/assets/images/USDC.png'),
+        sources: [
+          {
+            src: require('@/assets/images/USDC@2x.png'),
+            variant: '2x'
+          }
+        ]
+      } as PictureDescriptor,
 
-      token: undefined as TokenWithBalance | undefined,
-      amount: undefined as string | undefined,
-      nativeAmount: undefined as string | undefined,
+      //prepare
+      isLoading: true,
+      isProcessing: false,
+      isTokenSelectedByUser: false,
+      inputMode: 'TOKEN' as InputMode,
+      inputAsset: undefined as TokenWithBalance | undefined,
+      inputAmount: '',
+      inputAmountNative: '',
+      transferData: undefined as TransferData | undefined,
+      transferError: undefined as undefined | string,
+
+      //to tx
       isSubsidizedEnabled: false,
       estimatedGasCost: undefined as string | undefined,
-      transferData: undefined as TransferData | undefined,
       actionGasLimit: undefined as string | undefined,
       approveGasLimit: undefined as string | undefined
     };
   },
   computed: {
-    ...mapState('account', ['networkInfo', 'currentAddress', 'provider']),
-    hasBackButton(): boolean {
-      return this.txStep === undefined;
-    },
+    ...mapState('account', [
+      'networkInfo',
+      'currentAddress',
+      'nativeCurrency',
+      'gasPrices',
+      'ethPrice',
+      'tokens',
+      'savingsAPY',
+      'usdcPriceInWeth',
+      'savingsBalance',
+      'provider'
+    ]),
+    ...mapGetters('account', ['treasuryBonusNative']),
     outputUSDCAsset(): SmallTokenInfoWithIcon {
       return getUSDCAssetData(this.networkInfo.network);
     },
-    isNeedTransfer(): boolean {
-      if (this.token === undefined) {
+    nativeCurrencySymbol(): string {
+      return this.nativeCurrency.toUpperCase();
+    },
+    isSwapNeeded(): boolean {
+      if (this.inputAsset === undefined) {
         return true;
       }
 
-      return !sameAddress(this.token.address, this.outputUSDCAsset.address);
+      return !sameAddress(
+        this.inputAsset.address,
+        this.outputUSDCAsset.address
+      );
+    },
+    description(): string {
+      return (
+        this.isSwapNeeded
+          ? this.$t('savings.deposit.txtAssetWillBeConverted')
+          : this.$t('savings.txtUSDCCoinIsAStable')
+      ) as string;
+    },
+    hasBackButton(): boolean {
+      return this.step !== 'loader';
+    },
+    estimatedAnnualEarning(): string {
+      let possibleSavingsBalance = '0';
+      if (
+        this.inputAsset &&
+        sameAddress(this.inputAsset.address, this.outputUSDCAsset.address)
+      ) {
+        possibleSavingsBalance = this.inputAmount;
+      } else if (this.transferData !== undefined) {
+        possibleSavingsBalance = fromWei(
+          this.transferData.buyAmount,
+          this.outputUSDCAsset.decimals
+        );
+      }
+
+      if (this.savingsBalance !== undefined) {
+        possibleSavingsBalance = add(
+          this.savingsBalance,
+          possibleSavingsBalance
+        );
+      }
+
+      const usdcNative = multiply(this.usdcPriceInWeth, this.ethPrice);
+      const usdcAmountNative = multiply(possibleSavingsBalance, usdcNative);
+      const apyNative = multiply(
+        divide(this.savingsAPY, 100),
+        usdcAmountNative
+      );
+
+      return `~ $${formatToNative(apyNative)}`;
+    },
+    formattedUSDCTotal(): string {
+      if (this.inputAsset === undefined) {
+        return '0 USDC';
+      }
+
+      if (sameAddress(this.inputAsset.address, this.outputUSDCAsset.address)) {
+        return `${formatToNative(this.inputAmount)} USDC`;
+      }
+
+      if (this.transferData !== undefined) {
+        const boughtUSDC = fromWei(
+          this.transferData.buyAmount,
+          this.outputUSDCAsset.decimals
+        );
+        return `${formatToNative(boughtUSDC)} USDC`;
+      }
+
+      return '';
+    },
+    isNeedTransfer(): boolean {
+      if (this.inputAsset === undefined) {
+        return true;
+      }
+
+      return !sameAddress(
+        this.inputAsset.address,
+        this.outputUSDCAsset.address
+      );
+    }
+  },
+  watch: {
+    tokens: {
+      immediate: true,
+      handler(newVal: Array<TokenWithBalance>) {
+        try {
+          if (this.isTokenSelectedByUser) {
+            return;
+          }
+          const eth = newVal.find((t: TokenWithBalance) => t.address === 'eth');
+          if (eth) {
+            this.inputAsset = eth;
+          }
+        } finally {
+          this.isLoading = false;
+        }
+      }
     }
   },
   methods: {
+    ...mapActions('modals', { setModalIsDisplayed: 'setIsDisplayed' }),
     ...mapActions('account', {
       updateWalletAfterTxn: 'updateWalletAfterTxn'
     }),
     handleBack(): void {
-      if (this.isShowReview) {
-        this.isShowReview = !this.isShowReview;
+      if (this.step === 'review') {
+        this.step = 'prepare';
       } else {
         this.$router.back();
       }
     },
-    handleTxReview(args: {
-      token: TokenWithBalance;
-      amount: string;
-      nativeAmount: string;
-      isSubsidizedEnabled: boolean;
-      estimatedGasCost: string;
-      actionGasLimit: string;
-      approveGasLimit: string;
-      transferData?: TransferData;
-    }): void {
-      this.token = args.token;
-      this.amount = args.amount;
-      this.nativeAmount = args.nativeAmount;
-      this.transferData = args.transferData;
-      this.isSubsidizedEnabled = args.isSubsidizedEnabled;
-      this.estimatedGasCost = args.estimatedGasCost;
-      this.actionGasLimit = args.actionGasLimit;
-      this.approveGasLimit = args.approveGasLimit;
+    async handleTxReview(): Promise<void> {
+      if (this.inputAsset === undefined) {
+        return;
+      }
 
-      this.isShowReview = true;
+      this.isSubsidizedEnabled = false;
+      this.estimatedGasCost = undefined;
+      this.actionGasLimit = '0';
+      this.approveGasLimit = '0';
+      this.isProcessing = true;
+      try {
+        const gasLimits = await this.estimateAction(
+          this.inputAmount,
+          this.inputAsset,
+          this.transferData
+        );
+
+        this.actionGasLimit = gasLimits.actionGasLimit;
+        this.approveGasLimit = gasLimits.approveGasLimit;
+
+        console.info('Savings deposit action gaslimit:', this.actionGasLimit);
+        console.info('Savings deposit approve gaslimit:', this.approveGasLimit);
+
+        if (!isZero(this.actionGasLimit)) {
+          this.isSubsidizedEnabled = this.checkSubsidizedAvailability(
+            this.actionGasLimit
+          );
+          this.estimatedGasCost = this.subsidizedTxNativePrice(
+            this.actionGasLimit
+          );
+        }
+      } catch (err) {
+        this.isSubsidizedEnabled = false;
+        console.error(err);
+        Sentry.captureException("can't estimate savings deposit for subs");
+        return;
+      } finally {
+        this.isProcessing = false;
+      }
+
+      this.step = 'review';
+    },
+    subsidizedTxNativePrice(actionGasLimit: string): string | undefined {
+      const gasPrice = this.gasPrices?.FastGas.price ?? '0';
+      const ethPrice = this.ethPrice ?? '0';
+      if (isZero(gasPrice) || isZero(actionGasLimit) || isZero(ethPrice)) {
+        console.log(
+          "With empty parameter we can't calculate subsidized tx native price"
+        );
+        return undefined;
+      }
+      return calcTransactionFastNativePrice(
+        gasPrice,
+        actionGasLimit,
+        this.ethPrice
+      );
+    },
+    checkSubsidizedAvailability(actionGasLimit: string): boolean {
+      const gasPrice = this.gasPrices?.FastGas.price ?? '0';
+      const ethPrice = this.ethPrice ?? '0';
+      if (isZero(gasPrice) || isZero(actionGasLimit) || isZero(ethPrice)) {
+        console.log(
+          "With empty parameter we don't allow subsidized transaction"
+        );
+        return false;
+      }
+
+      if (this.inputAsset?.address === 'eth') {
+        console.info('Subsidizing for deposit ETH denied');
+        return false;
+      }
+
+      return isSubsidizedAllowed(
+        gasPrice,
+        actionGasLimit,
+        this.ethPrice,
+        this.treasuryBonusNative
+      );
+    },
+    async estimateAction(
+      inputAmount: string,
+      inputAsset: SmallToken,
+      transferData: TransferData | undefined
+    ): Promise<CompoundEstimateResponse> {
+      const resp = await estimateDepositCompound(
+        inputAsset,
+        this.outputUSDCAsset,
+        inputAmount,
+        transferData,
+        this.networkInfo.network,
+        this.provider.web3,
+        this.currentAddress
+      );
+      if (resp.error) {
+        this.transferError = this.$t('estimationError') as string;
+        Sentry.captureException("can't estimate savings deposit");
+        throw new Error(`Can't estimate action ${resp.error}`);
+      }
+      return resp;
+    },
+    async handleUpdateAmount(val: string): Promise<void> {
+      await this.updateAmount(val, this.inputMode);
+    },
+    async updateAmount(value: string, mode: InputMode): Promise<void> {
+      if (this.inputAsset === undefined || this.isLoading) {
+        return;
+      }
+
+      this.isLoading = true;
+
+      try {
+        if (!this.isSwapNeeded) {
+          console.log('Dont need transfer, token is USDC');
+          this.transferData = undefined;
+          if (mode === 'TOKEN') {
+            this.inputAmount = value;
+            this.inputAmountNative = convertNativeAmountFromAmount(
+              value,
+              this.inputAsset.priceUSD
+            );
+          } else {
+            this.inputAmount = convertAmountFromNativeValue(
+              value,
+              this.inputAsset.priceUSD,
+              this.inputAsset.decimals
+            );
+            this.inputAmountNative = value;
+          }
+        } else {
+          if (mode === 'TOKEN') {
+            this.inputAmount = value;
+            this.inputAmountNative = new BigNumber(
+              convertNativeAmountFromAmount(value, this.inputAsset.priceUSD)
+            ).toFixed(2);
+            const inputInWei = toWei(value, this.inputAsset.decimals);
+            this.transferData = await getTransferData(
+              this.outputUSDCAsset.address,
+              this.inputAsset.address,
+              inputInWei,
+              true,
+              '0.01',
+              this.networkInfo.network
+            );
+            this.transferError = undefined;
+          } else {
+            this.inputAmountNative = value;
+            const inputInWei = toWei(
+              convertAmountFromNativeValue(
+                value,
+                this.inputAsset.priceUSD,
+                this.inputAsset.decimals
+              ),
+              this.inputAsset.decimals
+            );
+            this.transferData = await getTransferData(
+              this.outputUSDCAsset.address,
+              this.inputAsset.address,
+              inputInWei,
+              true,
+              '0.01',
+              this.networkInfo.network
+            );
+            this.transferError = undefined;
+            this.inputAmount = fromWei(
+              this.transferData.sellAmount,
+              this.inputAsset.decimals
+            );
+          }
+        }
+      } catch (err) {
+        if (err instanceof ZeroXSwapError) {
+          this.transferError = mapError(err.publicMessage);
+        } else {
+          this.transferError = this.$t('exchangeError') as string;
+          Sentry.captureException(err);
+        }
+        console.error(`transfer error: ${err}`);
+        this.transferData = undefined;
+        if (mode === 'TOKEN') {
+          this.inputAmountNative = '0';
+        } else {
+          this.inputAmount = '0';
+        }
+      } finally {
+        this.isLoading = false;
+      }
+    },
+    async handleOpenSelectModal(): Promise<void> {
+      const token = await this.setModalIsDisplayed({
+        id: ModalType.SearchToken,
+        value: true,
+        payload: {
+          useWalletTokens: true
+        }
+      });
+
+      if (token === undefined) {
+        return;
+      }
+      this.isTokenSelectedByUser = true;
+      this.inputAsset = token;
+      this.transferData = undefined;
+      this.transferError = undefined;
+      this.inputAmount = '';
+      this.inputAmountNative = '';
+    },
+    handleToggleInputMode(): void {
+      if (this.inputMode === 'NATIVE') {
+        this.inputMode = 'TOKEN';
+        return;
+      }
+      this.inputMode = 'NATIVE';
+    },
+    async handleSelectMaxAmount(): Promise<void> {
+      if (this.inputAsset === undefined) {
+        return;
+      }
+      if (this.inputMode === 'TOKEN') {
+        await this.updateAmount(this.inputAsset.balance, 'TOKEN');
+      } else {
+        await this.updateAmount(
+          new BigNumber(
+            multiply(this.inputAsset.balance, this.inputAsset.priceUSD)
+          ).toFixed(2, BigNumber.ROUND_DOWN),
+          'NATIVE'
+        );
+      }
     },
     async handleTxStart(args: { isSmartTreasury: boolean }): Promise<void> {
-      if (this.token === undefined) {
-        console.error('token is empty during `handleTxStart`');
+      if (this.inputAsset === undefined) {
+        console.error('inputAsset is empty during `handleTxStart`');
         Sentry.captureException("can't start savings deposit TX");
         return;
       }
 
-      if (this.amount === undefined) {
-        console.error('amount is empty during `handleTxStart`');
+      if (this.inputAmount === '') {
+        console.error('inputAmount is empty during `handleTxStart`');
         Sentry.captureException("can't start savings deposit TX");
         return;
       }
@@ -141,27 +570,28 @@ export default Vue.extend({
 
       console.log('is smart treasury:', args.isSmartTreasury);
 
-      this.txStep = 'Confirm';
+      this.step = 'loader';
+      this.transactionStep = 'Confirm';
       try {
         await depositCompound(
-          this.token,
+          this.inputAsset,
           this.outputUSDCAsset,
-          this.amount,
+          this.inputAmount,
           this.transferData,
           this.networkInfo.network,
           this.provider.web3,
           this.currentAddress,
           args.isSmartTreasury,
           async () => {
-            this.txStep = 'Process';
+            this.transactionStep = 'Process';
           },
           this.actionGasLimit,
           this.approveGasLimit
         );
-        this.txStep = 'Success';
+        this.transactionStep = 'Success';
         this.updateWalletAfterTxn();
       } catch (err) {
-        this.txStep = 'Reverted';
+        this.transactionStep = 'Reverted';
         console.log('Savings deposit swap reverted');
         Sentry.captureException(err);
       }
