+++ src/views/savings/savings-withdraw-wrapper.vue
@@ -4,59 +4,175 @@
     hide-title
     @back="handleBack"
   >
-    <savings-withdraw-form v-if="!isShowReview" @tx-review="handleTxReview" />
-    <savings-withdraw-review
-      v-else-if="txStep === undefined"
-      :amount="amount"
+    <prepare-form
+      v-if="step === 'prepare'"
+      :asset="inputAsset"
+      :header-description="$t('savings.withdraw.txtWithdrawDescription')"
+      :header-title="$t('savings.withdraw.lblWithdrawFromSavings')"
+      :input-amount="inputAmountNative"
+      :input-amount-native="inputAmountNative"
+      :input-asset-heading="$t('savings.withdraw.lblWhatDoWeWithdraw')"
+      :input-mode="inputMode"
+      :is-loading="isLoading"
+      :is-processing="isProcessing"
+      :operation-description="$t('savings.withdraw.txtIfYouKeepSavings')"
+      :operation-title="estimatedAnnualEarning"
+      :output-asset-heading-text="$t('savings.deposit.lblAmountWeDepositIn')"
+      :selected-token-description="$t('savings.txtUSDCCoinIsAStable')"
+      :transfer-error="transferError"
+      @review-tx="handleTxReview"
+      @select-max-amount="handleSelectMaxAmount"
+      @update-amount="handleUpdateAmount"
+    />
+    <review-form
+      v-else-if="step === 'review'"
+      :amount="inputAmountNative"
+      :button-text="$t('savings.deposit.lblDepositInSavings')"
       :estimated-gas-cost="estimatedGasCost"
+      :header-title="$t('savings.deposit.lblReviewYourDeposit')"
+      :image="savings"
+      :input-amount-native-title="$t('savings.deposit.lblAndTotalOf')"
+      :input-amount-title="$t('savings.deposit.lblAmountWeDepositIn')"
       :is-subsidized-enabled="isSubsidizedEnabled"
-      :token="token"
+      :native-amount="inputAmountNative"
+      :token="inputAsset"
       @tx-start="handleTxStart"
     />
-    <full-page-form-loader v-else :step="txStep" />
+    <loader-form v-else-if="step === 'loader'" :step="transactionStep" />
   </secondary-page>
 </template>
 
 <script lang="ts">
 import Vue from 'vue';
-import { mapActions, mapState } from 'vuex';
+import { mapActions, mapGetters, mapState } from 'vuex';
 
 import * as Sentry from '@sentry/vue';
 
+import { TransferData } from '@/services/0x/api';
+import { divide, isZero, lessThan, multiply } from '@/utils/bigmath';
+import { formatToNative } from '@/utils/format';
 import { withdrawCompound } from '@/wallet/actions/savings/withdraw/withdraw';
-import { TokenWithBalance } from '@/wallet/types';
-
-import { FullPageFormLoader } from '@/components/controls/full-page-form-loader';
-import { SecondaryPage } from '@/components/layout/secondary-page';
 import {
-  SavingsWithdrawForm,
-  SavingsWithdrawReview
-} from '@/components/savings';
+  CompoundEstimateResponse,
+  estimateWithdrawCompound
+} from '@/wallet/actions/savings/withdraw/withdrawEstimate';
+import {
+  calcTransactionFastNativePrice,
+  isSubsidizedAllowed
+} from '@/wallet/actions/subsidized';
+import { getUSDCAssetData } from '@/wallet/references/data';
+import {
+  SmallToken,
+  SmallTokenInfoWithIcon,
+  TokenWithBalance
+} from '@/wallet/types';
+
+import { InputMode, LoaderForm, LoaderStep } from '@/components/forms';
+import PrepareForm from '@/components/forms/prepare-form/prepare-form.vue';
+import ReviewForm from '@/components/forms/review-form.vue';
+import { PictureDescriptor } from '@/components/html5';
+import { SecondaryPage } from '@/components/layout/secondary-page';
+
+type ProcessStep = 'prepare' | 'review' | 'loader';
 
 export default Vue.extend({
   name: 'SavingsWithdrawWrapper',
   components: {
-    FullPageFormLoader,
-    SavingsWithdrawReview,
-    SavingsWithdrawForm,
+    ReviewForm,
+    PrepareForm,
+    LoaderForm,
     SecondaryPage
   },
   data() {
     return {
-      isShowReview: false,
-      txStep: undefined as string | undefined,
+      //current
+      step: 'prepare' as ProcessStep,
+      transactionStep: undefined as LoaderStep | undefined,
+      savings: {
+        alt: this.$t('savings.lblSavings'),
+        src: require('@/assets/images/Savings@1x.png'),
+        sources: [
+          { src: require('@/assets/images/Savings@1x.png') },
+          {
+            variant: '2x',
+            src: require('@/assets/images/Savings@2x.png')
+          }
+        ],
+        webpSources: [
+          { src: require('@/assets/images/Savings@1x.webp') },
+          {
+            variant: '2x',
+            src: require('@/assets/images/Savings@2x.webp')
+          }
+        ]
+      } as PictureDescriptor,
+      usdcPicture: {
+        src: require('@/assets/images/USDC.png'),
+        sources: [
+          {
+            src: require('@/assets/images/USDC@2x.png'),
+            variant: '2x'
+          }
+        ]
+      } as PictureDescriptor,
 
-      token: undefined as TokenWithBalance | undefined,
-      amount: undefined as string | undefined,
+      //prepare
+      isLoading: false,
+      isProcessing: false,
+      isTokenSelectedByUser: false,
+      inputMode: 'NATIVE' as InputMode,
+      inputAmountNative: '',
+      transferData: undefined as TransferData | undefined,
+      transferError: undefined as undefined | string,
+
+      //to tx
       isSubsidizedEnabled: false,
       estimatedGasCost: undefined as string | undefined,
       actionGasLimit: undefined as string | undefined
     };
   },
   computed: {
-    ...mapState('account', ['networkInfo', 'currentAddress', 'provider']),
+    ...mapState('account', [
+      'networkInfo',
+      'currentAddress',
+      'savingsBalance',
+      'usdcPriceInWeth',
+      'savingsAPY',
+      'provider',
+      'ethPrice',
+      'gasPrices'
+    ]),
+    ...mapGetters('account', ['treasuryBonusNative', 'usdcNativePrice']),
     hasBackButton(): boolean {
-      return this.txStep === undefined;
+      return this.step !== 'loader';
+    },
+    USDCAsset(): SmallTokenInfoWithIcon {
+      return getUSDCAssetData(this.networkInfo.network);
+    },
+    inputAsset(): TokenWithBalance {
+      return {
+        address: this.USDCAsset.address,
+        decimals: this.USDCAsset.decimals,
+        symbol: this.USDCAsset.symbol,
+        name: 'USD Coin',
+        priceUSD: this.usdcNativePrice,
+        logo: this.USDCAsset.iconURL,
+        balance: this.savingsBalance,
+        marketCap: Number.MAX_SAFE_INTEGER
+      };
+    },
+    estimatedAnnualEarning(): string {
+      let possibleSavingsBalance = '0';
+
+      if (this.savingsBalance !== undefined) {
+        possibleSavingsBalance = this.savingsBalance;
+      }
+
+      const usdcNative = multiply(this.usdcPriceInWeth, this.ethPrice);
+      const usdcAmountNative = multiply(possibleSavingsBalance, usdcNative);
+      let apyNative = multiply(divide(this.savingsAPY, 100), usdcAmountNative);
+
+      return `~ $${formatToNative(apyNative)}`;
     }
   },
   methods: {
@@ -64,37 +180,103 @@ export default Vue.extend({
       updateWalletAfterTxn: 'updateWalletAfterTxn'
     }),
     handleBack(): void {
-      if (this.isShowReview) {
-        this.isShowReview = !this.isShowReview;
+      if (this.step === 'review') {
+        this.step = 'prepare';
       } else {
         this.$router.back();
       }
     },
-    handleTxReview(args: {
-      token: TokenWithBalance;
-      amount: string;
-      isSubsidizedEnabled: boolean;
-      estimatedGasCost: string;
-      actionGasLimit: string;
-    }): void {
-      console.log('TX REVIEW ');
-      this.token = args.token;
-      this.amount = args.amount;
-      this.isSubsidizedEnabled = args.isSubsidizedEnabled;
-      this.estimatedGasCost = args.estimatedGasCost;
-      this.actionGasLimit = args.actionGasLimit;
-
-      this.isShowReview = true;
+    async estimateAction(
+      amount: string,
+      asset: SmallToken
+    ): Promise<CompoundEstimateResponse> {
+      const resp = await estimateWithdrawCompound(
+        asset,
+        amount,
+        this.networkInfo.network,
+        this.provider.web3,
+        this.currentAddress
+      );
+      if (resp.error) {
+        Sentry.captureException("can't estimate savings deposit");
+        throw new Error(`Can't estimate action ${resp.error}`);
+      }
+      return resp;
     },
-    async handleTxStart(args: { isSmartTreasury: boolean }): Promise<void> {
-      if (this.token === undefined) {
-        console.error('token is empty during `handleTxStart`');
-        Sentry.captureException("can't start savings withdraw TX");
-        return;
+    subsidizedTxNativePrice(actionGasLimit: string): string | undefined {
+      const gasPrice = this.gasPrices?.FastGas.price ?? '0';
+      const ethPrice = this.ethPrice ?? '0';
+      if (isZero(gasPrice) || isZero(actionGasLimit) || isZero(ethPrice)) {
+        console.log(
+          "With empty parameter we can't calculate subsidized tx native price"
+        );
+        return undefined;
+      }
+      return calcTransactionFastNativePrice(
+        gasPrice,
+        actionGasLimit,
+        this.ethPrice
+      );
+    },
+    checkSubsidizedAvailability(actionGasLimit: string): boolean {
+      const gasPrice = this.gasPrices?.FastGas.price ?? '0';
+      const ethPrice = this.ethPrice ?? '0';
+      if (isZero(gasPrice) || isZero(actionGasLimit) || isZero(ethPrice)) {
+        console.log(
+          "With empty parameter we don't allow subsidized transaction"
+        );
+        return false;
       }
 
-      if (this.amount === undefined) {
-        console.error('amount is empty during `handleTxStart`');
+      return isSubsidizedAllowed(
+        gasPrice,
+        actionGasLimit,
+        this.ethPrice,
+        this.treasuryBonusNative
+      );
+    },
+    async handleTxReview(): Promise<void> {
+      this.isSubsidizedEnabled = false;
+      this.estimatedGasCost = undefined;
+      this.actionGasLimit = '0';
+      this.isProcessing = true;
+      try {
+        const gasLimits = await this.estimateAction(
+          this.inputAmountNative,
+          this.inputAsset
+        );
+
+        this.actionGasLimit = gasLimits.actionGasLimit;
+
+        console.info('Savings withdraw action gaslimit:', this.actionGasLimit);
+
+        if (!isZero(this.actionGasLimit)) {
+          this.isSubsidizedEnabled = this.checkSubsidizedAvailability(
+            this.actionGasLimit
+          );
+          this.estimatedGasCost = this.subsidizedTxNativePrice(
+            this.actionGasLimit
+          );
+        }
+      } catch (err) {
+        this.isSubsidizedEnabled = false;
+        console.error(err);
+        Sentry.captureException("can't estimate savings deposit for subs");
+      } finally {
+        this.isProcessing = false;
+      }
+
+      this.step = 'review';
+    },
+    handleSelectMaxAmount(): void {
+      this.inputAmountNative = this.inputAsset.balance;
+    },
+    handleUpdateAmount(val: string): void {
+      this.inputAmountNative = val;
+    },
+    async handleTxStart(args: { isSmartTreasury: boolean }): Promise<void> {
+      if (this.inputAmountNative === '') {
+        console.error('inputAmount is empty during `handleTxStart`');
         Sentry.captureException("can't start savings withdraw TX");
         return;
       }
@@ -107,24 +289,25 @@ export default Vue.extend({
 
       console.log('is smart treasury:', args.isSmartTreasury);
 
-      this.txStep = 'Confirm';
+      this.step = 'loader';
+      this.transactionStep = 'Confirm';
       try {
         await withdrawCompound(
-          this.token,
-          this.amount,
+          this.inputAsset,
+          this.inputAmountNative,
           this.networkInfo.network,
           this.provider.web3,
           this.currentAddress,
           this.actionGasLimit,
           args.isSmartTreasury,
           async () => {
-            this.txStep = 'Process';
+            this.transactionStep = 'Process';
           }
         );
-        this.txStep = 'Success';
+        this.transactionStep = 'Success';
         this.updateWalletAfterTxn();
       } catch (err) {
-        this.txStep = 'Reverted';
+        this.transactionStep = 'Reverted';
         console.log('Savings withdraw swap reverted');
         Sentry.captureException(err);
       }
