+++ src/wallet/references/data.ts
@@ -404,8 +404,6 @@ const getAssetsForTreasury = (
       address: move.address,
       decimals: move.decimals,
       symbol: move.symbol,
-      isFavorite: false,
-      isVerified: true,
       name: move.name,
       priceUSD: moveNativePrice,
       logo: move.iconURL,
@@ -415,8 +413,6 @@ const getAssetsForTreasury = (
       address: slp.address,
       decimals: slp.decimals,
       symbol: slp.symbol,
-      isFavorite: false,
-      isVerified: true,
       name: 'Sushi MOVE-ETH LP',
       priceUSD: slpNativePrice,
       logo: 'https://protocol-icons.s3.amazonaws.com/sushi-exchange.png',
